integrate transforms into vjepa2 model lazy import for models
